# Task/optimize centroid restraint