--- a/timemachine/cpp/src/centroid_restraint.hpp
+++ b/timemachine/cpp/src/centroid_restraint.hpp
@@ -13,6 +13,9 @@ private:
     int *d_group_a_idxs_;
     int *d_group_b_idxs_;
 
+    RealType * d_centroid_a_;
+    RealType * d_centroid_b_;
+
     int N_A_;
     int N_B_;