--- a/timemachine/cpp/src/kernels/k_centroid_restraint.cuh
+++ b/timemachine/cpp/src/kernels/k_centroid_restraint.cuh
@@ -2,88 +2,77 @@
 
 #include "k_fixed_point.cuh"
 
-__device__ const double PI = 3.14159265358979323846;
+template<typename RealType>
+void __global__ k_calc_centroid(
+    const double * __restrict__ d_coords,  // [n, 3]
+    const int * __restrict__ d_group_a_idxs,
+    const int * __restrict__ d_group_b_idxs,
+    const int N_A,
+    const int N_B,
+    RealType * d_centroid_a, // [3]
+    RealType * d_centroid_b // [3]
+    ){
+    int t_idx = blockDim.x*blockIdx.x + threadIdx.x;
+    if (N_A + N_B <= t_idx) {
+        return;
+    }
+    int group_idx = t_idx < N_A ? t_idx : t_idx - N_A;
+    RealType * centroid = t_idx < N_A ? d_centroid_a : d_centroid_b;
+    const int * cur_array = t_idx < N_A ? d_group_a_idxs : d_group_b_idxs;
+    for(int d=0; d < 3; d++) {
+        atomicAdd(centroid + d, d_coords[cur_array[group_idx]*3+d]);
+    }
+ }
 
 template<typename RealType>
 void __global__ k_centroid_restraint(
-    const int N,     // number of bonds
-    const double *coords,  // [n, 3]
-    const int *group_a_idxs,
-    const int *group_b_idxs,
+    // const int N,     // number of bonds, ignore for now
+    const double * __restrict__ d_coords,  // [n, 3]
+    const int * __restrict__ d_group_a_idxs,
+    const int * __restrict__ d_group_b_idxs,
     const int N_A,
     const int N_B,
-    // const double *masses, // ignore masses for now
+    const RealType * __restrict__ d_centroid_a, // [3]
+    const RealType * __restrict__ d_centroid_b, // [3]
+    // const double *d_masses, // ignore d_masses for now
     const double kb,
     const double b0,
-    unsigned long long *grad_coords,
-    unsigned long long *energy) {
+    unsigned long long *d_du_dx,
+    unsigned long long *d_u) {
 
-    // (ytz): ultra shitty inefficient algorithm, optimize later
-    const auto t_idx = blockDim.x*blockIdx.x + threadIdx.x;
-
-    if(t_idx != 0) {
+    const int t_idx = blockDim.x*blockIdx.x + threadIdx.x;
+    if (N_A + N_B <= t_idx) {
         return;
     }
-
-    double group_a_ctr[3] = {0};
-    for(int d=0; d < 3; d++) {
-        for(int i=0; i < N_A; i++) {
-            group_a_ctr[d] += coords[group_a_idxs[i]*3+d];
-        }
-        group_a_ctr[d] /= N_A;
+    int group_idx = t_idx < N_A ? t_idx : t_idx - N_A;
+    int count = t_idx < N_A ? N_A : N_B;
+    const int * cur_array = t_idx < N_A ? d_group_a_idxs : d_group_b_idxs;
+    RealType sign = t_idx < N_A ? 1.0 : -1.0;
+    RealType dij = 0;
+    RealType deltas[3];
+    for (int d=0; d < 3; d++) {
+        deltas[d] = d_centroid_a[d] / N_A - d_centroid_b[d] / N_B;
+        dij += deltas[d]*deltas[d];
     }
+    dij = sqrt(dij);
 
-    double group_b_ctr[3] = {0};
-    for(int d=0; d < 3; d++) {
-
-        for(int i=0; i < N_B; i++) {
-            group_b_ctr[d] += coords[group_b_idxs[i]*3+d];
-        }
-        group_b_ctr[d] /= N_B;
+    if(t_idx == 0 && d_u) {
+        RealType nrg = kb*(dij-b0)*(dij-b0);
+        d_u[0] = FLOAT_TO_FIXED<RealType>(nrg);
     }
-
-    double dx = group_a_ctr[0] - group_b_ctr[0];
-    double dy = group_a_ctr[1] - group_b_ctr[1];
-    double dz = group_a_ctr[2] - group_b_ctr[2];
-    double dij = sqrt(dx*dx + dy*dy + dz*dz);
-    double nrg = kb*(dij-b0)*(dij-b0);
-
-    if(energy) {
-        atomicAdd(energy, FLOAT_TO_FIXED<RealType>(nrg));
-    }
-
-
 
     // grads
-    if(grad_coords) {
+    if(d_du_dx) {
         for(int d=0; d < 3; d++) {
-
             if(b0 != 0) {
-
-                double du_ddij = 2*kb*(dij-b0);
-                double ddij_dxi = (group_a_ctr[d] - group_b_ctr[d])/dij;
-                for(int i=0; i < N_A; i++) {
-                    double delta = du_ddij*ddij_dxi/N_A;
-                    atomicAdd(grad_coords + group_a_idxs[i]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
-                }
-
-                for(int i=0; i < N_B; i++) {
-                    double delta = -du_ddij*ddij_dxi/N_B;
-                    atomicAdd(grad_coords + group_b_idxs[i]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
-                }
+                RealType du_ddij = 2*kb*(dij-b0);
+                RealType ddij_dxi = deltas[d]/dij;
+                RealType delta = sign*du_ddij*ddij_dxi/count;
+                atomicAdd(d_du_dx + cur_array[group_idx]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
             } else {
-                for(int i=0; i < N_A; i++) {
-                    double delta = 2*kb*(group_a_ctr[d] - group_b_ctr[d])/N_A;
-                    atomicAdd(grad_coords + group_a_idxs[i]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
-                }
-
-                for(int i=0; i < N_B; i++) {
-                    double delta = -2*kb*(group_a_ctr[d] - group_b_ctr[d])/N_B;
-                    atomicAdd(grad_coords + group_b_idxs[i]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
-                }
+                RealType delta = sign*2*kb*deltas[d]/count;
+                atomicAdd(d_du_dx + cur_array[group_idx]*3 + d, FLOAT_TO_FIXED<RealType>(delta));
             }
-
         }
     }
-
 }